+++ src/apps/settings/modules/Home/MiniStore.tsx
@@ -1,26 +1,64 @@
+import { Resource } from '@seelen-ui/lib/types';
+import { ResourceText } from '@shared/components/ResourceText';
 import { Skeleton } from 'antd';
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
+import { useTranslation } from 'react-i18next';
 
 import cs from './MiniStore.module.css';
 
-export function ProductSkeleton() {
+export function ResourceSkeleton() {
   return (
-    <div className={cs.product}>
-      <Skeleton.Image active />
+    <div className={cs.resourceSkeleton}>
+      <Skeleton.Image active style={{ width: '100%', aspectRatio: '1 / 1' }} />
       <Skeleton active paragraph={false} />
     </div>
   );
 }
 
-export function MiniStore() {
-  const [products, _setProducts] = useState<any[]>([]);
+interface Featured {
+  newArrivals: Resource[];
+  top: Resource[];
+  popular: Resource[];
+  staffLiked: Resource[];
+}
+
+export function RemoteResources() {
+  const [resources, setResources] = useState<Resource[]>([]);
+
+  const { t } = useTranslation();
+
+  useEffect(() => {
+    fetch('https://product.seelen.io/resources/featured')
+      .then((res) => res.json())
+      .then((data: Featured) => setResources(data.top))
+      .catch(() => {});
+  }, []);
 
   return (
     <>
-      <h1 className={cs.title}>New Resources</h1>
-      <div className={cs.miniStore}>
-        {products.length === 0 &&
-          Array.from({ length: 10 }).map((_, i) => <ProductSkeleton key={i} />)}
+      <h1 className={cs.title}>{t('home.new_resources')}</h1>
+      <div className={cs.resources}>
+        {resources.length === 0 &&
+          Array.from({ length: 10 }).map((_, i) => <ResourceSkeleton key={i} />)}
+
+        {resources.map((resource) => {
+          if (!resource.metadata.portrait) return null;
+          return (
+            <a
+              key={resource._id}
+              href={`https://seelen.io/resources/${resource.friendlyId.replace('@', '')}`}
+              target="_blank"
+              className={cs.resource}
+            >
+              <img src={resource.metadata.portrait} />
+              <div className={cs.text}>
+                <span>
+                  <ResourceText text={resource.metadata.displayName} />
+                </span>
+              </div>
+            </a>
+          );
+        })}
       </div>
     </>
   );