+++ src/apps/settings/modules/Home/index.tsx
@@ -1,11 +1,11 @@
-import { MiniStore } from './MiniStore';
+import { RemoteResources } from './MiniStore';
 import { NoticeSlider } from './News';
 
 export function Home() {
   return (
     <>
       <NoticeSlider />
-      <MiniStore />
+      <RemoteResources />
     </>
   );
 }